requirements.txt
PyGithub
gitpython
requests
openai         # optional, if using OpenAI
transformers   # optional, for local models
flask          # or fastapi
pytest
flake8
bandit
semgrep        # optional
radon
